User roles and role validation while deleting file

+++ convex/files.ts
@@ -31,7 +31,7 @@ async function hasAccessToOrg(ctx: QueryCtx | MutationCtx, orgId: string) {
   if (!user) return null;
 
   const hasAccess =
-    user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId);
+    user.orgIds.some(item => item.orgId === orgId) || user.tokenIdentifier.includes(orgId);
 
   if (!hasAccess) return null;
 
@@ -111,6 +111,12 @@ export const deleteFile = mutation({
 
     if (!access) {
       throw new ConvexError("no access to file");
+    }
+
+    const isAdmin = access.user.orgIds.find(org => org.orgId === access.file.orgId)?.role === 'admin'
+
+    if(!isAdmin) {
+      throw new ConvexError("You do not have access to delete this file")
     }
 
     await ctx.db.delete(args.fileId);
